Moved automations service to separate file
towards https://linear.app/ghost/issue/NY-1286
ref https://github.com/TryGhost/Ghost/pull/28120

This change should have no impact on functionality.

We have [a lint rule that limits the length of `index.js` files][0]. In
[an upcoming change][1], the automations service will cross that limit.

This patch moves it into a separate file. Doing so also makes it a
little easier to test.

[0]: https://github.com/TryGhost/eslint-plugin-ghost/blob/22d154a8e3807e4e30219fb8449e7aa9b90a110b/lib/config/node.js#L32-L37
[1]: https://github.com/TryGhost/Ghost/pull/28120
